(base) ☁  coauthor [refactoring/prepare-py] ⚡  git diff --cached | ask cm -m copilot-claude-3.5-sonnet refactor: split AgentConfig into separate module from agent_dataclass
- Improves code organization by separating configuration concerns
- Updates all import references across multiple files to use new module path
- Maintains existing functionality while improving code structure
