+++ coauthor/agent/agent_base.py
@@ -26,7 +26,8 @@ from ..utils.repetition import check_for_massive_repetition
 from ..utils.xml import get_xml_format_from_files
 
 # Local imports - agent components
-from .agent_dataclass import AgentConfig, AgentSettings, AgentPrompts
+from .agent_config import AgentConfig
+from .agent_dataclass import AgentSettings, AgentPrompts
 from .agent_state import AgentStateRound, AgentStateGlobal
 from .tool_state import ToolState
 from .logdb import create_log_entry, update_log_statistics, update_log_output_files
